CI: replace Travis with GH Actions
This translates the existing Travis workflow to a GH Actions workflow.

Effectively it is the same workflow.

At a detail level, there are the following differences:
* The commands to download Phive have been removed in favour of letting the `setup-php` action install it automatically.
* A build against PHP 8.0 has been added which is not allowed to fail.
* A build against PHP 8.1 has been added, which is (allowed to fail).

The build against PHP 8.1 also _does_ currently fail, but this is due to a PHP 8.1 incompatibility in Phive.
This issue has been reported in https://github.com/phar-io/phive/issues/319 and :fingers_crossed: https://github.com/phar-io/phive/pull/320 should fix it.